% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnBias.R
\name{plotAnBias}
\alias{plotAnBias}
\title{Plot An bias statistics}
\usage{
plotAnBias(expData, probeData, plotNamePrefix = NULL)
}
\arguments{
\item{expData}{- result of the procCustArray function}

\item{probeData}{- data.frame with detailed probe information}

\item{plotNamePrefix}{- prefix used to name all output plots}
}
\description{
Plot An bias statistics
}
