% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotG4Bias.R
\name{plotG4Bias}
\alias{plotG4Bias}
\title{Plot G4 bias statistics}
\usage{
plotG4Bias(expData, probeData, plotNamePrefix = NULL)
}
\arguments{
\item{expData}{- result of the procCustArray function}

\item{probeData}{- data.frame with detailed probe information}

\item{plotNamePrefix}{- prefix used to name all output plots}
}
\description{
Plot G4 bias statistics
}
