% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGCcontentBias.R
\name{plotGCcontentBias}
\alias{plotGCcontentBias}
\title{Plot GC-content bias statistics}
\usage{
plotGCcontentBias(expData, plotName = NULL)
}
\arguments{
\item{expData}{- result of the procCustArray function}

\item{plotName}{- name of the output plot}
}
\description{
Plot GC-content bias statistics
}
