% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLengthBias.R
\name{plotLengthBias}
\alias{plotLengthBias}
\title{Plot length bias statistics}
\usage{
plotLengthBias(expData, plotName = NULL)
}
\arguments{
\item{expData}{- result of the procCustArray function}

\item{plotName}{- name of the output plot}
}
\description{
Plot length bias statistics
}
