% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotT7Bias.R
\name{plotT7Bias}
\alias{plotT7Bias}
\title{Plot T7 bias statistics}
\usage{
plotT7Bias(expData, probeData, plotNamePrefix = NULL)
}
\arguments{
\item{expData}{- result of the procCustArray function}

\item{probeData}{- data.frame with detailed probe information}

\item{plotNamePrefix}{- prefix used to name all output plots}
}
\description{
Plot T7 bias statistics
}
