% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probediff.R
\name{probediff}
\alias{probediff}
\title{Compare probes from two distinct groups}
\usage{
probediff(
  ProbeData,
  SampleIDs,
  label1,
  label2,
  label1_desc,
  label2_desc,
  FilePrefix,
  FeatureValLabel,
  SampleID,
  ExpCut = 500,
  Version = 1
)
}
\arguments{
\item{ProbeData}{probe data obtained in the}

\item{SampleIDs}{ids of all samples in the dataset}

\item{label1}{label of the first compared feature}

\item{label2}{label of the second compared feature}

\item{label1_desc}{description of the first compared feature}

\item{label2_desc}{description of the second compared feature}

\item{FilePrefix}{prefix of the output files}

\item{FeatureValLabel}{description of the shown feature values used to label the plot axis}

\item{SampleID}{ID of the sample which will be used to plot detailes sample-level data}

\item{ExpCut}{expression level cutoff (only probes with signals above this value will be used)}

\item{Version}{version of the result files}
}
\description{
Compare probes from two distinct groups
}
