% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procCustArray.R
\name{procCustArray}
\alias{procCustArray}
\title{Preprocess data from custom Agilent microarray}
\usage{
procCustArray(targets_file, path)
}
\arguments{
\item{targets_file}{- table specifying association between samples and data files, for details see limma::readTargets function description}

\item{path}{- directory containing the targets file and all data files}
}
\description{
Preprocess data from custom Agilent microarray
}
